fix: prevent cleaning events when autoDelTimeAway is set to 0

+++ server/utils/common/event.ts
@@ -69,6 +69,7 @@ export class Event {
   }
   static async cleanEvents() {
     const conf = await this.getConf()
+    if (conf.autoDelTimeAway === 0) return
     const del = await prisma.event.deleteMany({
       where: {
         ctime: {
